fix(lint): clear the default clippy::all warnings across the workspace
First cargo clippy --workspace --all-targets run this codebase has ever
had (no CI runs clippy today). Baseline was a modest 46 warnings across
~20 real source locations once target/lib/bin/test duplicates are
collapsed -- fixed all of them except three deliberately left as-is:

- collapsible_if / collapsible_match: merged nested ifs into a single
  && condition, or into a match guard, everywhere the inner block had
  no else and the match's fallback arm was a no-op `_ => {}` (verified
  per-site -- a guard-collapse changes behavior if the fallback arm does
  something, so this was checked, not applied blindly).
- clippy::single_match: match-with-one-arm-plus-`_=>{}` in receipt.rs
  rewritten as `if let`.
- clippy::unnecessary_sort_by (5 sites): sort_by(|a,b| b.f.cmp(&a.f)) ->
  sort_by_key(|x| Reverse(x.f)); all sort keys are Copy primitives, sort
  stability unaffected (both are stable sorts).
- clippy::double_ended_iterator_last, explicit_counter_loop,
  unnecessary_cast, needless_bool_assign, unused_imports (x2): mechanical,
  zero behavior change.
- clippy::wildcard_in_or_patterns in receipt.rs's --format match: kept
  the named "terminal"|"ansi"|"receipt" arms as documentation of the
  supported values (an unrecognized format intentionally falls back to
  the terminal receipt rather than erroring) and suppressed the lint
  explicitly with a reason comment, rather than deleting the names or
  silently changing the fallback to an error.

Deliberately NOT fixed, flagged in the session report instead:
- Two clippy::too_many_arguments sites (TrajectoryChunk::new at 9/7
  args across 12 call sites; run_usage_command at 8/7 args, 1 call
  site) -- real API-ergonomics smells, but fixing means a signature
  change (builder or params struct), which is a design call, not a
  mechanical lint fix.
- clippy::manual_checked_ops in pr_blame.rs: already guarded against
  the div-by-zero it's warning about; checked_div wouldn't be more
  correct, just a different spelling of the same guard.
- clippy::write_literal in two adapter test fixtures (cline.rs,
  windsurf.rs): cosmetic, and the fix means hand-escaping embedded
  JSON/braces in a raw string literal -- not worth the transcription
  risk in test fixture data for a pure style warning.

One nested-if collapse in blunder.rs initially left a stray extra
closing brace (caught by self-review, confirmed by lenovo's compiler
before this was ever pushed anywhere).

Verified on lenovo: cargo build --workspace exit 0, cargo test
--workspace exit 0, 345 passed / 0 failed / 0 ignored.

+++ apps/cli/src/commands/pr_blame.rs
@@ -96,7 +96,7 @@ pub fn annotate_pr_files(
                 primary_outcome: outcome,
                 outcome_confidence: conf,
                 spend_usd: spend,
-                total_tokens: Some(first.total_tokens as u64),
+                total_tokens: Some(first.total_tokens),
             });
         } else {
             annotations.push(FileAiProvenance {
